pyproject.toml
[project]
name = "agent-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain[openai]>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.2.1",
    "langgraph>=1.0.9",
    "pypdf>=6.7.4",
    "langgraph-cli[inmem]>=0.4.12",
    "deepagents>=0.4.4",
    "langchain-deepseek>=1.0.1",
    "fastapi>=0.133.1",
    "uvicorn[standard]>=0.41.0",
    "httpx>=0.28.1",
    "requests>=2.32.5",
    "python-dotenv>=1.2.1",
    "langgraph-checkpoint-postgres>=3.0.4",
    "psycopg[binary]>=3.3.3",
    "langchain-openai>=1.1.10",
]
